requirements.txt
torch>=2.0.0
numpy>=1.22
PyYAML>=6.0
pandas>=1.3
scikit-learn>=1.0
scipy>=1.8
matplotlib>=3.5
opencv-python>=4.7.0
torchaudio>=2.0.0
transformers>=4.31

